[REFACTOR] User ID Long 타입 전환 및 Timezone(Instant) 전체 적용

+++ backend/src/main/java/com/example/pproject/application/controller/JobApplicationController.java
@@ -26,7 +26,7 @@ public class JobApplicationController {
     @PostMapping
     public ResponseEntity<Long> apply(@RequestBody @Valid JobApplicationRequest request,
                                       @AuthenticationPrincipal JwtUserPrincipal user) {
-        Long applicationId = jobApplicationService.apply(request, Integer.valueOf(user.getUserid()));
+        Long applicationId = jobApplicationService.apply(request, Long.valueOf(user.getUserid()));
         return ResponseEntity.ok(applicationId);
     }
